Added MapFileReader
